ok update

+++ cmds/moderación/clear.js
@@ -4,20 +4,19 @@ module.exports = {
         alias: ['purge'], //Alias
         description: "Borrar mensajes", //Descripción (OPCIONAL)
         usage: "z!clear numero",
-        category: 'moderacion'
+        category: 'moderacion',
+        botPermissions: ['MANAGE_MESSAGES'],
+        memberPermissions: ['MANAGE_MESSAGES']
 
     }, run: async ({ message, args, embedResponse }) => {
 
-        if (!message.member.hasPermission('MANAGE_MESSAGES')) return embedResponse('No tienes el permiso `MANAGE_MESSAGES`!')
-        if (!message.guild.me.hasPermission('MANAGE_MESSAGES')) return embedResponse('No tengo el permiso `MANAGE_MESSAGES`')
-        if (!args[0]) return embedResponse('Escribe un numero!')
-        if (isNaN(args[0])) return embedResponse('Escribe un numero!')
-        if (args[0] >= 100 || args[0] === 0) return embedResponse('Un numero del 1 al 99')
+        if (isNaN(args[0]) || !args[0]) return embedResponse('<:cancel:779536630041280522> | Escribe la cantidad de mensajes que quieres borrar.')
+        if (args[0] >= 100 || args[0] <= 0) return embedResponse('<:cancel:779536630041280522> | Escribe un numero entre el 1 al 100')
         await message.delete()
         await message.channel.bulkDelete(args[0], true).then(d => {
-            if (d.size < args[0]) return d.size === 0 ? embedResponse('Ningun mensaje fue eliminado!') : embedResponse('Mensajes eliminados: ' + d.size)
+            if (d.size < args[0]) return d.size === 0 ? embedResponse('<:cancel:779536630041280522> | Ningun mensaje fue eliminado.') : embedResponse('<:accept:779536642365063189> | Mensajes eliminados: ' + d.size)
 
-            else return embedResponse('Mensajes eliminados: ' + d.size)
+            else return embedResponse('<:accept:779536642365063189> | Mensajes eliminados: ' + d.size)
 
         })
     }
